feat: start-intro-screen
