+++ app/src/main/AndroidManifest.xml
@@ -11,9 +11,10 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.MiniProject_CarRacing"
+
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".StartActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -22,5 +23,6 @@
             </intent-filter>
         </activity>
     </application>
+    <uses-permission android:name="android.permission.POST_NOTIFICATIONS" />
 
 </manifest>
