+++ app/src/main/res/layout/activity_start.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/background"
-    tools:context=".MainActivity">
+    tools:context=".StartActivity">
 
     <!-- Tên game -->
     <TextView
@@ -16,8 +16,8 @@
         android:layout_marginTop="92dp"
         android:text="RACING FURY"
         android:textColor="#FF3B3B"
-        android:textFontWeight="1000"
-        android:textSize="40dp"
+        android:textStyle="bold"
+        android:textSize="40sp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
