--- app/src/main/res/raw/click_sound.mp3
@@ -1,10 +1,0 @@
-btnVolume.setOnClickListener(v -> {
-    if (isPlaying) {
-        mediaPlayer.pause();
-        Toast.makeText(this, "Tạm dừng nhạc", Toast.LENGTH_SHORT).show();
-    } else {
-        mediaPlayer.start();
-        Toast.makeText(this, "Phát nhạc", Toast.LENGTH_SHORT).show();
-    }
-    isPlaying = !isPlaying;
-});
