+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <string name="app_name">MiniProject_CarRacing</string>
+    <string name="app_name">MiniProject Car Racing</string>
+
+    <string name="intro_title">Chào mừng đến với cuộc đua!</string>
+    <string name="intro_description">Tham gia vào cuộc đua hấp dẫn, chọn chế độ chơi và bắt đầu hành trình tốc độ của bạn ngay bây giờ!</string>
+    <string name="start_race">Bắt đầu đua</string>
 </resources>
